requirements.txt
imageio==2.28.1
matplotlib==3.7.5
numpy==1.23.5
opencv_contrib_python==4.8.0.76
opencv_python==4.9.0.80
opencv_python_headless==4.8.1.78
Pillow==10.4.0
PyYAML==6.0.1
setuptools==45.2.0
skimage==0.0
timm==0.9.2
torch==2.1.0
torchvision==0.16.0
tqdm==4.65.0